Ravitheja-Reddy/requirements.txt
# Licensed under the MIT License
streamlit
requests
PyMuPDF
pytesseract
Pillow
